working anthropic function call
